+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/services/FileService.java
@@ -6,6 +6,7 @@ import com.udacity.jwdnd.course1.cloudstorage.model.User;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.naming.AuthenticationException;
 import java.io.IOException;
 import java.util.List;
 
@@ -24,8 +25,13 @@ public class FileService {
         return fileMapper.getDuplicatedFiles(user.getUserId(), fileName).isEmpty();
     }
 
-    public File getFile(Integer fileId) {
-        return fileMapper.getFile(fileId);
+    public File getFile(Integer fileId, String username) throws AuthenticationException {
+        File file = fileMapper.getFile(fileId);
+        User user = userService.getUser(username);
+        if (user.getUserId() != file.getUserId()) {
+            throw new AuthenticationException("You do not have access to this file.");
+        }
+        return file;
     }
 
     public List<File> getAllFiles(String username) {
@@ -35,14 +41,21 @@ public class FileService {
 
     public Integer uploadFile(MultipartFile mFile, String username) throws IOException {
         User user = userService.getUser(username);
-        File file = new File(null, mFile.getOriginalFilename(),
-                mFile.getContentType(), mFile.getSize(), user.getUserId(), mFile.getBytes());
-
-        return this.fileMapper.uploadFile(file);
+        try {
+            File file = new File(null, mFile.getOriginalFilename(),
+                    mFile.getContentType(), mFile.getSize(), user.getUserId(), mFile.getBytes());
+            return this.fileMapper.uploadFile(file);
+        } catch (Exception e) {
+            throw e;
+        }
     }
 
-    public Integer deleteFile(File file, String username) {
+    public Integer deleteFile(File file, String username) throws AuthenticationException {
         User user = userService.getUser(username);
+        File existingFile = fileMapper.getFile(file.getFileId());
+        if (user.getUserId() != existingFile.getUserId()) {
+            throw new AuthenticationException("You do not have access to this file.");
+        }
         file.setUserId(user.getUserId());
         return fileMapper.deleteFile(file.getFileId());
     }
